+++ testdata/app/helpers/admin/reports_helper.rb
@@ -1,0 +1,11 @@
+module Admin
+  module ReportsHelper
+    def report_title(report)
+      report.title
+    end
+
+    def report_period(report)
+      "#{report.start_date} - #{report.end_date}"
+    end
+  end
+end